+++ src/onegov.org/onegov/org/app.py
@@ -325,6 +325,7 @@ def get_check_contrast_asset():
 def get_code_editor_asset():
     yield 'ace.js'
     yield 'ace-mode-form.js'
+    yield 'ace-mode-markdown.js'
     yield 'ace-mode-xml.js'
     yield 'ace-mode-yaml.js'
     yield 'ace-theme-tomorrow.js'